test: extract ornament auto test model
